requirements (1).txt
appdirs==1.4.4
audioread==3.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
decorator==5.1.1
ffmpeg-python==0.2.0
filelock==3.9.0
Flask==2.2.3
future==0.18.3
huggingface-hub==0.12.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lazy_loader==0.1
librosa==0.10.0
llvmlite==0.39.1
MarkupSafe==2.1.2
more-itertools==9.0.0
msgpack==1.0.4
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
openai-whisper==20230124
packaging==23.0
pandas==1.5.3
pooch==1.6.0
pycparser==2.21
pydub==0.25.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.3
textblob==0.17.1
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
